Add Cut 2
Finished the functionality of the Add Cut button so that it works separately for each of the cells and deletes upon a second click. The only issue is that if you don't select the start of a line, the line is too short. Still need to make it so that if there's already a cut line in the cell, it asks if you want to replace it or create a new event.